Sync develop branch changes March 4 - 6 to hdf5_1_14 branch (#4121)
* Do not enable szip for sanitizer runs (#4057)

* Add note to H5Tset_fields about needing to set datatype precision first (#4059)

* Offset of a floating-point type also needs to be accounted for

* Clarify ordering of H5Tset_precision and H5Tset_fields

* Fix issue where H5Tset_fields does not account for datatype offsets (#4061)

H5Tset_fields did not account for any offset in a floating-point datatype,
causing it to fail when a datatype's precision is correctly set such that
it doesn't include the offset bits.

* Ignore UserPresets and Use only C compiler for sanitizers (#4066)

* Remove user presets file

* Only use C compiler for sanitzers

* Rename incorrectly named option (#4067)

* Rename incorrectly named option

* Restore the correct uses of USING_MEMCHECKER

* Update release note

* Fix a memory leak in the cmpd_dset test (#4071)

This was due to not freeing a test buffer. It was not a core
library memory leak.

* Fix uninitialized bytes in cmpd_dset test (#4072)

Compound fill values were set to the integer -1, causing valgrind
to flag 'uninitialized bytes' errors.

This is just a problem with the cmpd_dset test and not a core
library problem.

* Update INSTALL files (#4052)

* Add NEWSLETTER and merge abi reports and add sha256sums (#4055)

* Fix uninitialized bytes in selection I/O test (#4073)

This was due to a complex type fill value being set to -1 instead
of a proper complex value. This was a test problem and not a core
library issue.

* fix path for S3 build path in CI (#4076)

* Correct paths for 1.14 and add lines missing from release_docs/INSTALL_CMake.txt.

---------

Co-authored-by: Allen Byrne <50328838+byrnHDF@users.noreply.github.com>
Co-authored-by: jhendersonHDF <jhenderson@hdfgroup.org>
Co-authored-by: Dana Robinson <43805+derobins@users.noreply.github.com>