Remove ``-lm -ldl`` from hdf5.pc on macOS
Motivation
----------

Prior to this commit, the Libs.private entry within hdf5.pc contain the
``-lm`` and ``-ldl`` flags on macOS. Becuase macOS (for at least a
number of recent releases) doesn't specify actually contain these
libraries, this means that passing the flags produced by

   ``pkg-config --static --libs hdf5``

will produce linker errors on macOS. The contents of these particular
libraries are instead all consolidated within a single system library.

The Solution
------------

It turns out that this issue originates from some "magic" that CMake
does to removes these flags from all calls to the linker on macOS. This
"magic" seems to cause the built-in logic for checking for the existence
of these libraries to falsely report that the libraries exist on macOS.
These false positives propagate to the contents of the contents of
hdf5.pc. At the same time, this same "magic" seems to prevent
linker-errors during the actual compilation of hdf5.

Concerns
--------

I think that much older versions of macOS may have actually provided the
"m" and "dl" libraries. If that is the case, then my fix may break
builds on these versions of macOS... I'm not really sure when this
transition occured.