Removed h5test.* from the C++ test
Revised the C++ tests to eliminate the use of private header files. The solution is to use some functions in the tools library and duplicate some functions from h5test.*.  Some of these duplicated functions can be added to the tools library at a later time, however.

Also, there are many small changes to improve the overall outputs of the tests.

Note: The changes to c++/test/Makefile.in will be committed soon.

(Issue GH-4517)