Add CMake build mode flags to libhdf5.settings
The build mode flags for things like optimization are not added
to CMAKE_C_FLAGS and were not exported to the libhdf5.settings
file. This change corrects that.

Also removes -Og from the developer mode flags since that's already
set via the debug flags the developer mode is based on.