Switch H5FD interface and all VFDs to us (H5OPEN <id>) idiom
Move all VFDs into the 'H5FD' package, instead of their own individual
packages.

Stop using VFD's 'term' callback to reset the ID for the VFD.  Use explicit
'register' and 'unregister' calls to do that.  Use the 'term' callback only
for dealing with VFD internal resource cleanup.

A _ton_ of other misc. changes, but mainly to update and align the VFDs more
with current best practices in the library.

Signed-off-by: Quincey Koziol <quincey@koziol.cc>