Remove version number from h5repack test plugin
The h5repack test uses an older copy of a plugin from test that
appended the version number to group names. The plugin in test was
updated to use a fixed string instead, but the h5repack copy wasn't
updated.

We'll remove this copy and just use the plugin from test after
the Autotools are removed.