Update stdio API calls
The stdio VFD set up ftell, etc. equivalents differently than the
rest of the library. This PR aligns them with H5private.h and
H5win32defs.h and should fix some issues with MinGW. The stdio VFD
now also uses HDoff_t, which will be a public symbol in the next
major release of HDF5.

In the main library, HDftell was set to be ftell and not ftello.
This probably isn't a big deal since the systems where it's used
all have 64-bit longs, but we should match types correctly.