Fix clang enum-enum-conversion warnings
clang complains that H5FD_MEM_NTYPES, which is a member of H5FD_mem_t,
is being used with a variable of type H5F_mem_page_t. This is fine
since H5F_mem_t is a subset of H5F(D)_mem_page_t and we're restricting
ourselves to the subset in this case.