Fix reading bad size in the raw header continuation message
This issue was reported in GH-5376 as a heap-use-after-free vulnerability in
one of the free lists.  It appeared that the library came to this vulnerability
after it encountered an undetected reading of a bad value.  The fuzzer now failed
with an appropriate error message.

This considers addressing what GH-5376 reported.