Use ptrdiff_t for ssize_t on Windows
The handstands we do to support ssize_t on Windows are a little
baroque. Other libraries just use ptrdiff_t since it's signed
and almost certainly the same size in bytes as size_t.

The PR adds a compile assert to ensure that size_t and ssize_t
have the same size, just to be sure.