Set max POSIX IO size to 0x7ffff000 bytes
This was previously set to SSIZE_MAX on Linux, which is correct as far
as POSIX is concerned, but Linux defines the max as MAX_RW_COUNT, which
is set to INT_MAX & PAGE_CACHE_MASK (0x7ffff000) in fs.h. Since this
is very close to the INT_MAX value that was previously set for MacOS
and Windows, HDF5 will now use the Linux value on all platforms.

Larger I/O sizes will be split into multiple smaller read/write calls
in the VFDs.

Fixes #5873