Restrict fscanf %s in h5import
A couple of places in the h5import.c code use fscanf with %s to read
strings into a fixed-size buffer without restricting the number
of characters, which could lead to a stack buffer overflow.

This fix restricts the number of characters that can be read to
the size of the buffer.